Update 1:
Structure of the website has been set:
About
Portfolio - 2 halves
Blog
Contact.

Portfolio and Contact will be lightboxes that appear on top of the webpage upon clicking on different sections.

The top navbar is just for the logo for a clean simple look, the menu to the side (only in large format) will aid in scrolling to different sections